+++ ExposurePlayback/Context/Components/HLSNative+ExposureContext+Creation.swift
@@ -22,6 +22,7 @@ extension Player where Tech == HLSNative<ExposureContext> {
     /// - parameter sessionToken: Token identifying this session
     public convenience init(environment: Environment, sessionToken: SessionToken) {
         self.init(environment: environment, sessionToken: sessionToken, analytics: ExposureAnalytics.self)
+
     }
     
     /// Creates and configures `Player` for use with `HLSNative` and `ExposureContext`.
@@ -29,8 +30,8 @@ extension Player where Tech == HLSNative<ExposureContext> {
     /// - parameter environment: The *Exposure* environment
     /// - parameter sessionToken: Token identifying this session
     /// - parameter analytics: The *Exposure* related `AnalyticsProvider` tasked with delivering analytics to the *EMP* backend.
-    public convenience init<Analytics: ExposureStreamingAnalyticsProvider>(environment: Environment, sessionToken: SessionToken, analytics: Analytics.Type) {
-        let generator: (Tech.Context.Source?) -> AnalyticsProvider = { _ in return analytics.init(environment: environment, sessionToken: sessionToken) }
+    public convenience init<Analytics: ExposureStreamingAnalyticsProvider>(environment: Environment, sessionToken: SessionToken, analytics: Analytics.Type, cdn: CDNInfoFromEntitlement? = nil , analyticsFromPlayback: AnalyticsFromEntitlement? = nil ) {
+        let generator: (Tech.Context.Source?) -> AnalyticsProvider = { _ in return analytics.init(environment: environment, sessionToken: sessionToken, cdn: cdn, analytics: analyticsFromPlayback) }
         let context = ExposureContext(environment: environment, sessionToken: sessionToken)
         let tech = HLSNative<ExposureContext>()
         tech.airplayHandler = context