+++ ExposurePlayback/Events/BitrateChanged.swift
@@ -20,10 +20,16 @@ extension Playback {
         /// New bitrate, in kilobit/s
         internal let bitrate: Int64
         
-        internal init(timestamp: Int64, offsetTime: Int64?, bitrate: Int64) {
+        internal var cdnInfo: CDNInfoFromEntitlement?
+        
+        internal var analyticsInfo: AnalyticsFromEntitlement?
+        
+        internal init(timestamp: Int64, offsetTime: Int64?, bitrate: Int64, cdnInfo: CDNInfoFromEntitlement? = nil , analyticsInfo: AnalyticsFromEntitlement? = nil) {
             self.timestamp = timestamp
             self.offsetTime = offsetTime
             self.bitrate = bitrate
+            self.cdnInfo = cdnInfo
+            self.analyticsInfo = analyticsInfo
         }
     }
 }
@@ -48,6 +54,20 @@ extension Playback.BitrateChanged: AnalyticsEvent {
             json[JSONKeys.offsetTime.rawValue] = value
         }
         
+        if let cdnInfo = cdnInfo {
+            json[JSONKeys.profile.rawValue] = cdnInfo.profile
+            json[JSONKeys.host.rawValue] = cdnInfo.host
+            json[JSONKeys.provider.rawValue] = cdnInfo.provider
+        }
+        
+        if let analyticsInfo = analyticsInfo {
+            json[JSONKeys.bucket.rawValue] = analyticsInfo.bucket
+            json[JSONKeys.postInterval.rawValue] = analyticsInfo.postInterval
+            json[JSONKeys.tag.rawValue] = analyticsInfo.tag
+        }
+        
+        json[JSONKeys.StreamingTechnology.rawValue] = "HLS"
+        
         return json
     }
     
@@ -56,5 +76,17 @@ extension Playback.BitrateChanged: AnalyticsEvent {
         case timestamp = "Timestamp"
         case offsetTime = "OffsetTime"
         case bitrate = "Bitrate"
+        
+        // CDN
+        case profile = "profile"
+        case host = "host"
+        case provider = "provider"
+        
+        // Analytics info from entitlement
+        case bucket = "bucket"
+        case postInterval = "postInterval"
+        case tag = "tag"
+        
+        case StreamingTechnology = "StreamingTechnology"
     }
 }