+++ ExposurePlayback/Events/Completed.swift
@@ -18,9 +18,15 @@ extension Playback {
         /// Offset in the video sequence where the playback was stopped. This would typically be equal to the length of the asset in milliseconds.
         internal let offsetTime: Int64?
         
-        internal init(timestamp: Int64, offsetTime: Int64?) {
+        internal var cdnInfo: CDNInfoFromEntitlement?
+        
+        internal var analyticsInfo: AnalyticsFromEntitlement?
+        
+        internal init(timestamp: Int64, offsetTime: Int64?, cdnInfo: CDNInfoFromEntitlement? = nil , analyticsInfo: AnalyticsFromEntitlement? = nil) {
             self.timestamp = timestamp
             self.offsetTime = offsetTime
+            self.cdnInfo = cdnInfo
+            self.analyticsInfo = analyticsInfo
         }
     }
 }
@@ -44,6 +50,20 @@ extension Playback.Completed: AnalyticsEvent {
             json[JSONKeys.offsetTime.rawValue] = value
         }
         
+        if let cdnInfo = cdnInfo {
+            json[JSONKeys.profile.rawValue] = cdnInfo.profile
+            json[JSONKeys.host.rawValue] = cdnInfo.host
+            json[JSONKeys.provider.rawValue] = cdnInfo.provider
+        }
+        
+        if let analyticsInfo = analyticsInfo {
+            json[JSONKeys.bucket.rawValue] = analyticsInfo.bucket
+            json[JSONKeys.postInterval.rawValue] = analyticsInfo.postInterval
+            json[JSONKeys.tag.rawValue] = analyticsInfo.tag
+        }
+        
+        json[JSONKeys.StreamingTechnology.rawValue] = "HLS"
+        
         return json
     }
     
@@ -51,5 +71,17 @@ extension Playback.Completed: AnalyticsEvent {
         case eventType = "EventType"
         case timestamp = "Timestamp"
         case offsetTime = "OffsetTime"
+        
+        // CDN
+        case profile = "profile"
+        case host = "host"
+        case provider = "provider"
+        
+        // Analytics info from entitlement
+        case bucket = "bucket"
+        case postInterval = "postInterval"
+        case tag = "tag"
+        
+        case StreamingTechnology = "StreamingTechnology"
     }
 }