+++ ExposurePlayback/Events/ProgramChanged.swift
@@ -23,11 +23,18 @@ extension Playback {
         /// Length in milliseconds of this program, according to the EPG
         let videoLength: Int64?
         
-        internal init(timestamp: Int64, offsetTime: Int64?, programId: String, videoLength: Int64? = nil) {
+        internal var cdnInfo: CDNInfoFromEntitlement?
+        
+        internal var analyticsInfo: AnalyticsFromEntitlement?
+        
+        internal init(timestamp: Int64, offsetTime: Int64?, programId: String, videoLength: Int64? = nil, cdnInfo: CDNInfoFromEntitlement? = nil , analyticsInfo: AnalyticsFromEntitlement? = nil) {
             self.timestamp = timestamp
             self.offsetTime = offsetTime
             self.programId = programId
             self.videoLength = videoLength
+            
+            self.cdnInfo = cdnInfo
+            self.analyticsInfo = analyticsInfo
         }
     }
 }
@@ -56,6 +63,20 @@ extension Playback.ProgramChanged: AnalyticsEvent {
             json[JSONKeys.videoLength.rawValue] = videoLength
         }
         
+        if let cdnInfo = cdnInfo {
+            json[JSONKeys.profile.rawValue] = cdnInfo.profile
+            json[JSONKeys.host.rawValue] = cdnInfo.host
+            json[JSONKeys.provider.rawValue] = cdnInfo.provider
+        }
+        
+        if let analyticsInfo = analyticsInfo {
+            json[JSONKeys.bucket.rawValue] = analyticsInfo.bucket
+            json[JSONKeys.postInterval.rawValue] = analyticsInfo.postInterval
+            json[JSONKeys.tag.rawValue] = analyticsInfo.tag
+        }
+        
+        json[JSONKeys.StreamingTechnology.rawValue] = "HLS"
+        
         return json
     }
     
@@ -65,6 +86,18 @@ extension Playback.ProgramChanged: AnalyticsEvent {
         case offsetTime = "OffsetTime"
         case programId = "ProgramId"
         case videoLength = "VideoLength"
+        
+        // CDN
+        case profile = "profile"
+        case host = "host"
+        case provider = "provider"
+        
+        // Analytics info from entitlement
+        case bucket = "bucket"
+        case postInterval = "postInterval"
+        case tag = "tag"
+        
+        case StreamingTechnology = "StreamingTechnology"
     }
 }