+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # CHANGELOG
 
+* `2.2.60` Release - [2.2.60](#2260)
 * `2.2.51` Release - [2.2.51](#2251)
 * `2.2.50` Release - [2.2.50](#2250)
 * `2.2.40` Release - [2.2.40](#2240)
@@ -23,6 +24,14 @@
 * `2.0.79` Release - [2.0.79](#2079)
 * `2.0.78` Release - [2.0.78](#2078)
 * `0.77.x` Releases - [0.77.0](#0770)
+
+
+## 2.2.600
+#### Features
+* `EMP-15925` Analytics Enhancements 
+* `EMP-15925` Now developers can disable analytics when playing content.
+* `EMP-15983` use `PlaybackProperties` when using offline playback 
+
 
 ## 2.2.510
 #### Changes